+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 * [UINV-138](https://issues.folio.org/browse/UINV-138) Align actions icons in table to right hand side of view pane(s)
 * [UIOR-540](https://issues.folio.org/browse/UIOR-540) Add loading indicator when selecting new records
 
+### Bug fixes
+* [UIOR-562](https://issues.folio.org/browse/UIOR-562) Fix PO number validation error message not showing to user
+
 ## [2.0.4](https://github.com/folio-org/ui-orders/tree/v2.0.4) (2020-04-24)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v2.0.3...v2.0.4)